remove ext traits
