+++ src/buf/mod.rs
@@ -18,13 +18,24 @@
 
 mod buf_impl;
 mod buf_mut;
-pub mod ext;
+mod chain;
 mod iter;
+mod limit;
+#[cfg(feature = "std")]
+mod reader;
+mod take;
 mod vec_deque;
+#[cfg(feature = "std")]
+mod writer;
 
 pub use self::buf_impl::Buf;
 pub use self::buf_mut::BufMut;
+pub use self::chain::Chain;
 #[cfg(feature = "std")]
 pub use self::buf_mut::IoSliceMut;
-pub use self::ext::{BufExt, BufMutExt};
 pub use self::iter::IntoIter;
+pub use self::limit::Limit;
+pub use self::take::Take;
+
+#[cfg(feature = "std")]
+pub use self::{reader::Reader, writer::Writer};
