+++ tests/test_chain.rs
@@ -1,6 +1,5 @@
 #![warn(rust_2018_idioms)]
 
-use bytes::buf::{BufExt, BufMutExt};
 use bytes::{Buf, BufMut, Bytes};
 #[cfg(feature = "std")]
 use std::io::IoSlice;
